Start to refactor so that the tantivy index interaction is isolated from the http server code, and rework error handling a bit
