+++ Cargo.toml
@@ -18,3 +18,4 @@ serde_derive = "1.0"
 ammonia = "2"
 maplit = "1"
 itertools = "0.8"
+failure = "0.1"
