+++ src/server.rs
@@ -1,14 +1,9 @@
-use tantivy::collector::{Count, TopDocs};
-use tantivy::query::QueryParserError;
-use tantivy::query::QueryParserError::*;
+use tantivy::collector::{TopDocs};
+
 use tantivy::query::TermQuery;
 use tantivy::schema::IndexRecordOption;
-use tantivy::schema::Value::Facet;
-use tantivy::schema::Field;
+
 use ttvy::TantivyReader;
-use BookMetadata;
-//use tantivy::schema::Facet;
-use tantivy::schema::Schema;
 use tantivy::schema::Term;
 
 use epub::doc::EpubDoc;
@@ -17,6 +12,7 @@ use std::io;
 use std::fmt;
 use std::fs::File;
 use std::io::prelude::*;
+use error::ClientError;
 
 pub struct Server {
 	pub reader: TantivyReader,
@@ -24,18 +20,6 @@ pub struct Server {
 	pub port: u32,
 	pub use_coverdir: bool,
 	pub coverdir: Option<String>,
-	id_field: Field,
-	file_field: Field,
-	creator_field: Field,
-	mime_field: Field,
-	tags_field: Field,
-	title_field: Field,
-}
-
-#[derive(Debug, Serialize)]
-struct ClientError {
-	name: String,
-	msg: String,
 }
 
 #[derive(Debug)]
@@ -53,28 +37,13 @@ impl fmt::Display for ServerError {
 
 impl Server {
 	pub fn new(reader:TantivyReader,host:&str,port:u32, use_coverdir:bool, coverdir:Option<String>) -> Result<Server, ServerError> {
-		let searcher = reader.reader.searcher();
-		let schema = searcher.schema();
 		Ok(Server {
 			reader,
 			host:host.to_string(),
 			port,
 			use_coverdir,
 			coverdir,
-			id_field:Server::get_field(schema, "id")?,
-			file_field:Server::get_field(schema, "file")?,
-			creator_field:Server::get_field(schema, "creator")?,
-			mime_field:Server::get_field(schema, "cover_mime")?,
-			tags_field:Server::get_field(schema, "tags")?,
-			title_field:Server::get_field(schema, "title")?,
 		})
-	}
-
-	pub fn get_field(schema:&Schema, field_name:&str) -> Result<Field, ServerError> {
-		match schema.get_field(field_name) {
-			Some(f) => Ok(f),
-			None => Err(ServerError { msg: format!("Mismatching schema - specified field {} does not exist in tantivy schema for this index.", field_name).to_string() }),
-		}
 	}
 
 	#[allow(unreachable_code)]
@@ -85,64 +54,25 @@ impl Server {
 			rouille::log(&request, io::stdout(), || {
 				router!(request,
 					(GET) (/api/search) => {
-						let searcher = &self.reader.reader.searcher();
 						let q = &request.get_param("query");
 						let query_str = match q {
 							Some(query) => query,
 							None => return self.get_str_error_response("Query error", "\"query\" should be provided when performing a query")
 						}.trim();
 
-						let query_parsed = &self.reader.query_parser.parse_query(query_str);
-						let query = match query_parsed {
-							Err(e) => return self.get_error_response(&self.get_query_error(&e)),
-							Ok(q) => q,
-						};
-
-						//in theory we could cache the searcher for subsequent queries with a higher startat.. but too complex for now
-						let start_pos = match request.get_param("start").unwrap_or_else(|| "0".to_string()).parse::<usize>() {
-							Ok(pos) => pos,
+						let start = match request.get_param("start").unwrap_or_else(|| "0".to_string()).parse::<usize>() {
+							Ok(start) => start,
 							Err(_) => return self.get_str_error_response("Type error", "\"start\" should have an integer argument"),
 						};
 
-						let top_collector = TopDocs::with_limit(match request.get_param("limit").unwrap_or_else(|| "20".to_string()).parse::<usize>() {
+						let limit = match request.get_param("limit").unwrap_or_else(|| "20".to_string()).parse::<usize>() {
 							Ok(lim) => lim,
 							Err(_) => return self.get_str_error_response("Type error", "\"limit\" should have an integer argument"),
-						} + start_pos);
-
-						let count_collector = Count;
-
-						let docs = match searcher.search(query, &(top_collector, count_collector)) {
-							Ok(docs) => docs,
-							Err(e) => {println!("Error searching:{}", e); return self.get_str_error_response("Index error", "Something is wrong with the index")},
-						};
-
-
-						let num_docs = docs.0.len();
-						//json encode query value
-						let mut json_str: String = format!("{{\"count\":{}, \"position\":{}, \"query\":\"{}\", \"books\":[", docs.1, start_pos, query_str.replace("\"","\\\"")).to_owned();
-						for (i,doc) in docs.0.iter().enumerate() {
-							if (i+1)>start_pos {
-								let retrieved = match searcher.doc(doc.1) {
-									Ok(doc) => doc,
-									Err(_) => continue,
-								};
-
-								json_str.push_str(match &serde_json::to_string(&self.to_bm(&retrieved, &searcher.schema())) {
-									Ok(str) => str,
-									Err(_) => continue,
-								});
-
-								if (i+1)!=num_docs {
-									json_str.push_str(",");
-								}
-							}
 						}
-						json_str.push_str("]}");
 
 						rouille::Response::from_data("application/json", json_str).with_additional_header("Access-Control-Allow-Origin", "*")
 					},
 					(GET) (/api/book/{id: i64}) => {
-						//FIXME so many unwraps, damn
 						let searcher = &self.reader.reader.searcher();
 
 						let id_field = self.id_field;
@@ -152,7 +82,7 @@ impl Server {
 
 						//could this be better with TopFieldCollector which uses a FAST field?
 						let docs = searcher.search(&term_query, &TopDocs::with_limit(1)).unwrap();
-
+						
 						if docs.len()==1 {
 							let retrieved = searcher.doc(docs.first().unwrap().1.to_owned()).unwrap();
 							let file_loc = retrieved.get_first(self.file_field).unwrap().text().unwrap();
@@ -216,100 +146,6 @@ impl Server {
 				)
 			})
 		})
-	}
-
-	pub fn to_bm(&self, doc: &tantivy::Document, schema: &Schema) -> BookMetadata {
-		BookMetadata {
-			id: self.get_doc_i64("id", &doc, &schema), //not populated ?
-			title: self.get_doc_str("title", &doc, &schema),
-			description: self.get_doc_str("description", &doc, &schema),
-			publisher: self.get_doc_str("publisher", &doc, &schema),
-			creator: self.get_doc_str("creator", &doc, &schema),
-			subject: self.get_tags("tags", &doc),
-			file: self.get_doc_str("file", &doc, &schema).unwrap(),
-			filesize: self.get_doc_i64("filesize", &doc, &schema),
-			modtime: self.get_doc_i64("modtime", &doc, &schema),
-			pubdate: self.get_doc_str("pubdate", &doc, &schema),
-			moddate: self.get_doc_str("moddate", &doc, &schema),
-			cover_mime: self.get_doc_str("cover_mime", &doc, &schema),
-		}
-	}
-
-	//I *know* the fields are present in schema, and I *know* that certain fields eg id are always populated, so just unwrap() here
-	fn get_doc_str(&self, field: &str, doc: &tantivy::Document, schema: &Schema) -> Option<String> {
-		doc.get_first(schema.get_field(field).unwrap()).map(|val| match val.text() {
-			Some(t) => t.to_string(),
-			_ => "".to_string(),
-		})
-	}
-
-	fn get_doc_i64(&self, field: &str, doc: &tantivy::Document, schema: &Schema) -> i64 {
-		doc.get_first(schema.get_field(field).unwrap()).unwrap().i64_value()
-	}
-
-	fn get_tags(&self, _field: &str, doc: &tantivy::Document) -> Option<Vec<String>> {
-		let vals: Vec<&tantivy::schema::Value> = doc.get_all(self.tags_field);
-		if vals.is_empty() {
-			return None;
-		}
-		let mut tags = Vec::new();
-
-		for v in vals {
-			tags.push(
-				match v {
-					Facet(f) => f.encoded_str(),
-					_ => "",
-				}
-				.to_string(),
-			)
-		}
-
-		Some(tags)
-	}
-
-	fn get_query_error(&self, query_err: &QueryParserError) -> ClientError {
-		match query_err {
-			SyntaxError => ClientError {
-				name: "Syntax Error".to_string(),
-				msg: "There was a syntax error in the search string.".to_string(),
-			},
-			FieldDoesNotExist(s) => ClientError {
-				name: "Field does not Exist".to_string(),
-				msg: format!("You searched for a field that does not exist:{}", s),
-			},
-			ExpectedInt(_) => ClientError {
-				name: "Expected Integer".to_string(),
-				msg: "Search argument requires an integer.".to_string(),
-			},
-			AllButQueryForbidden => ClientError {
-				name: "All But query forbidden".to_string(),
-				msg: "Queries that only exclude (eg. \"-king\") are forbidden.".to_string(),
-			},
-			NoDefaultFieldDeclared => ClientError {
-				name: "No field declared".to_string(),
-				msg: "You must specify a field to query.".to_string(),
-			},
-			FieldNotIndexed(s) => ClientError {
-				name: "Field unknown".to_string(),
-				msg: format!("The field you searched for is unknown:{}", s),
-			},
-			UnknownTokenizer(_, _) => ClientError {
-				name: "Unknown Tokenizer".to_string(),
-				msg: "The tokenizer for the given field is unknown".to_string(),
-			},
-			FieldDoesNotHavePositionsIndexed(s) => ClientError {
-				name: "Field does not have positions indexed".to_string(),
-				msg: format!("Field does not have positions indexed: {}", s),
-			},
-			RangeMustNotHavePhrase => ClientError {
-				name: "Range must not have phrase".to_string(),
-				msg: "Range must not have phrase".to_string(),
-			},
-			DateFormatError(_) => ClientError {
-				name: "Date must have correct format".to_string(),
-				msg: "Date must have correct format".to_string(),
-			},
-		}
 	}
 
 	fn get_error_response(&self, client_error: &ClientError) -> rouille::Response {
